+++ src/main/java/com/master/caps/Controller/CourseLecturerController.java
@@ -31,7 +31,7 @@ public class CourseLecturerController {
 
     // 获取单个课程讲师关联
     @GetMapping("/{id}")
-    public ResponseEntity<CourseLecturer> getCourseLecturerById(@PathVariable Long id) {
+    public ResponseEntity<CourseLecturer> getCourseLecturerById(@PathVariable Integer id) {
         Optional<CourseLecturer> optionalCourseLecturer = courseLecturerRepository.findById(id);
         if (optionalCourseLecturer.isPresent()) {
             CourseLecturer courseLecturer = optionalCourseLecturer.get();
@@ -50,7 +50,7 @@ public class CourseLecturerController {
 
     // 更新课程讲师关联
     @PutMapping("/{id}")
-    public ResponseEntity<CourseLecturer> updateCourseLecturer(@PathVariable Long id, @RequestBody CourseLecturer courseLecturer) {
+    public ResponseEntity<CourseLecturer> updateCourseLecturer(@PathVariable Integer id, @RequestBody CourseLecturer courseLecturer) {
         Optional<CourseLecturer> optionalCourseLecturer = courseLecturerRepository.findById(id);
         if (optionalCourseLecturer.isPresent()) {
             CourseLecturer existingCourseLecturer = optionalCourseLecturer.get();
@@ -65,7 +65,7 @@ public class CourseLecturerController {
 
     // 删除课程讲师关联
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCourseLecturer(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteCourseLecturer(@PathVariable Integer id) {
         Optional<CourseLecturer> optionalCourseLecturer = courseLecturerRepository.findById(id);
         if (optionalCourseLecturer.isPresent()) {
             courseLecturerRepository.delete(optionalCourseLecturer.get());
