+++ src/main/java/com/master/caps/Controller/CourseScheduleController.java
@@ -30,7 +30,7 @@ public class CourseScheduleController {
 
     // 获取单个课程安排
     @GetMapping("/{id}")
-    public ResponseEntity<CourseSchedule> getCourseScheduleById(@PathVariable Long id) {
+    public ResponseEntity<CourseSchedule> getCourseScheduleById(@PathVariable Integer id) {
         Optional<CourseSchedule> optionalCourseSchedule = courseScheduleRepository.findById(id);
         if (optionalCourseSchedule.isPresent()) {
             CourseSchedule courseSchedule = optionalCourseSchedule.get();
@@ -49,7 +49,7 @@ public class CourseScheduleController {
 
     // 更新课程安排
     @PutMapping("/{id}")
-    public ResponseEntity<CourseSchedule> updateCourseSchedule(@PathVariable Long id, @RequestBody CourseSchedule courseSchedule) {
+    public ResponseEntity<CourseSchedule> updateCourseSchedule(@PathVariable Integer id, @RequestBody CourseSchedule courseSchedule) {
         Optional<CourseSchedule> optionalCourseSchedule = courseScheduleRepository.findById(id);
         if (optionalCourseSchedule.isPresent()) {
             CourseSchedule existingCourseSchedule = optionalCourseSchedule.get();
@@ -65,7 +65,7 @@ public class CourseScheduleController {
 
     // 删除课程安排
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCourseSchedule(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteCourseSchedule(@PathVariable Integer id) {
         Optional<CourseSchedule> optionalCourseSchedule = courseScheduleRepository.findById(id);
         if (optionalCourseSchedule.isPresent()) {
             courseScheduleRepository.deleteById(id);
