+++ src/main/java/com/master/caps/Controller/FacultyController.java
@@ -21,7 +21,7 @@ public class FacultyController {
     }
 
     @GetMapping("/{id}")
-    public Faculty getFacultyById(@PathVariable Long id) throws Exception {
+    public Faculty getFacultyById(@PathVariable Integer id) throws Exception {
         return facultyRepository.findById(id)
                 .orElseThrow(() -> new Exception("Faculty not found with id: " + id));
     }
@@ -32,7 +32,7 @@ public class FacultyController {
     }
 
     @PutMapping("/{id}")
-    public Faculty updateFaculty(@PathVariable Long id, @RequestBody Faculty updatedFaculty) throws Exception {
+    public Faculty updateFaculty(@PathVariable Integer id, @RequestBody Faculty updatedFaculty) throws Exception {
         return facultyRepository.findById(id)
                 .map(faculty -> {
                     faculty.setFacultyname(updatedFaculty.getFacultyname());
@@ -42,7 +42,7 @@ public class FacultyController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteFaculty(@PathVariable Long id) {
+    public void deleteFaculty(@PathVariable Integer id) {
         facultyRepository.deleteById(id);
     }
 }
